backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
requests
python-multipart
python-dotenv
ultralytics
opencv-python
pillow
numpy
pandas
ollama
